Added burst support into upload thunk and refactor runUploadThunk
